Changed Light Id from Composite Key to generated long Id
This was done to provide better separation between the different parts of the app.
The old LightID Object contained endpoint specific information that shouldn't
be in the ui layer.
This changes the Database, but doesn't provide a room migration, so that the app has to be reinstalled and a simple update doesn't work