Fix room database entity update.
After the transition from a composite key to an auto-generated primary key in commit 24fc00bb9299537623aa4a7ebceda6ab07624e31 all SQL UPDATE-statements for light objects failed silently. Only the first insert was successful, subsequent updates were not possible. This meant that toggling the light on via GUI was not persisted in the app's room database and therefore not shown when accessing the light again.
All UPDATE SQL-statements failed because room identifies entities to be updated via its primary key lightId. The lightId, however, is not known when parsed from the remote endpoint (network). This fix adds the following three cases to the upsert (update or insert) logic:
- Case 1: Insert if neither lightId (primary key) nor (endpointId and endpointLightId) are found inside the database.
- Case 2: Update if lightId (primary key) is given in the light object. A light with this lightId must already exist inside the database.
- Case 3: Update if endpointId and endpointLightId are given in the light object. A light with this endpointId and endpointLightId must already exist inside the database.