Improve logging (#92)
This improves the logging messages and moves some messages to verbose or info level.
The goal of this change is, that it should be possible to read the Log with debug Level output get enough information to know what the app is doing. But also not so much, that the relevant info is hidden between other output.