Improve Formatting rules (#102)
The previous formatting rules behaved sometimes a little bit strange with newlines. This should change that.

All files have been formatted with the new rules.

Android Studio still behaves a bit strange with format on save or when a file is reformatted by shortcut. But formatting on commit or formatting of the whole project works as expected.