Implement getChangePayload() for RecyclerView.
If the payloads list is not empty, the ViewHolder is currently bound to old data and Adapter may run an efficient partial update using the payload info. If the payload is empty, Adapter must run a full bind.
The payloads list is constructed by overwriting getChangePayload() from DiffUtil.ItemCallback. By using a partial bind, we can prevent the light card from "flickering". This is the default animation for updates.