Split app into 3 modules (#143)
This splts the app into 3 modules: app(frontend), backend and util.
The app module contains all frontend code and is the only module that can create the real app. This could also be renamed to frontend at a later date.
The backend module contains the Repos and everything that is behind them. The util module contains all utils that were in the util package and the serviceLocator.

This is only the split and does not yet make use of the different opportunites that come with such a new architecture, that can be done in further commits.

---------

Co-authored-by: Kai-Uwe Lankenau <github@familie-lankenau.de>