update Dependabot search path for gradle files (#185)
This adds the new modules util and backend to folders where dependabot
searches for dependencies.
It also adds the app module for completeness, even through this module
was already found by dependabot.
This also ignores Updates for gradle, because most times these updates
also need changes in gradle-wrapper.properties which is not handled by
dependabot